requirements.txt
amqp==2.3.2
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.4
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.1
django-bootstrap3==11.0.0
django-easy-pdf==0.1.1
django-scheduler==0.8.8
docraptor==1.2.0
flower==0.9.2
html5lib==1.0.1
icalendar==4.0.2
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
kombu==4.2.1
parso==0.3.1
pdfrw==0.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.9
simplegeneric==0.8.1
six==1.11.0
tinycss2==0.6.1
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.3
gunicorn